Move more UI things out of /http
