+++ http.js
@@ -31,31 +31,8 @@ function start (entry, opts) {
   var emitter = new EventEmitter()
   var id = 0
   var state = {
-    count: compiler.metadata.count,
-    files: {},
-    sse: 0,
-    size: 0
+    sse: 0
   }
-
-  files.forEach(function (filename) {
-    state.files[filename] = {
-      name: filename,
-      progress: 0,
-      timestamp: '        ',
-      size: 0,
-      status: 'pending',
-      done: false
-    }
-  })
-
-  compiler.on('error', function (topic, sub, err) {
-    if (err.pretty) state.error = err.pretty
-    else state.error = `${topic}:${sub} ${err.message}\n${err.stack}`
-  })
-
-  compiler.on('progress', function () {
-    state.error = null
-  })
 
   compiler.on('ssr', function (result) {
     state.ssr = result
@@ -64,16 +41,6 @@ function start (entry, opts) {
   compiler.on('change', function (nodeName, edgeName, nodeState) {
     var node = nodeState[nodeName][edgeName]
     var name = nodeName + ':' + edgeName
-    var data = {
-      name: nodeName,
-      progress: 100,
-      timestamp: time(),
-      size: 0,
-      status: 'done',
-      done: true
-    }
-    state.files[nodeName] = data
-
     if (name === 'documents:index.html') emitter.emit('documents:index.html', node)
     if (name === 'styles:bundle') emitter.emit('styles:bundle', node)
   })
@@ -164,7 +131,7 @@ function start (entry, opts) {
     emitter.on('documents:index.html', reloadScript)
     emitter.on('styles:bundle', reloadStyle)
     state.sse += 1
-    // if (!quiet) render()
+    compiler.emit('sse-connect')
 
     res.writeHead(200, {
       'Content-Type': 'text/event-stream',
@@ -190,7 +157,7 @@ function start (entry, opts) {
         emitter.removeListener('styles:bundle', reloadStyle)
         connected = false
         state.sse -= 1
-        // if (!quiet) render()
+        compiler.emit('sse-disconnect')
       }
     }
 
@@ -258,17 +225,4 @@ function gzip (buffer, req, res) {
   var zipper = gzipMaybe(req, res)
   pump(zipper, res)
   zipper.end(buffer)
-}
-
-function time () {
-  var date = new Date()
-  var hours = numPad(date.getHours())
-  var minutes = numPad(date.getMinutes())
-  var seconds = numPad(date.getSeconds())
-  return `${hours}:${minutes}:${seconds}`
-}
-
-function numPad (num) {
-  if (num < 10) num = '0' + num
-  return num
 }
