+++ lib/ui.js
@@ -1,7 +1,8 @@
 var ansi = require('ansi-escape-sequences')
-var differ = require('ansi-diff')
 var pretty = require('prettier-bytes')
+var gzipSize = require('gzip-size')
 var keypress = require('keypress')
+var differ = require('ansi-diff')
 var strip = require('strip-ansi')
 var nanoraf = require('nanoraf')
 
@@ -25,11 +26,63 @@ module.exports = createUi
 function createUi (compiler, state) {
   var diff = differ()
 
+  Object.assign(state, {
+    count: compiler.metadata.count,
+    files: {},
+    size: 0
+  })
+
   var render = nanoraf(onrender, raf)
 
-  compiler.on('error', render)
-  compiler.on('progress', render)
-  compiler.on('change', render)
+  files.forEach(function (filename) {
+    state.files[filename] = {
+      name: filename,
+      progress: 0,
+      timestamp: '        ',
+      size: 0,
+      status: 'pending',
+      done: false
+    }
+  })
+
+  compiler.on('error', function (topic, sub, err) {
+    if (err.pretty) state.error = err.pretty
+    else state.error = `${topic}:${sub} ${err.message}\n${err.stack}`
+    render()
+  })
+
+  compiler.on('progress', function () {
+    state.error = null
+    render()
+  })
+
+  compiler.on('change', function (nodeName, edgeName, nodeState) {
+    var node = nodeState[nodeName][edgeName]
+    var name = nodeName + ':' + edgeName
+    var data = {
+      name: nodeName,
+      progress: 100,
+      timestamp: time(),
+      size: 0,
+      status: 'done',
+      done: true
+    }
+    state.files[nodeName] = data
+
+    // Only calculate the gzip size if there's a buffer. Apparently zipping
+    // an empty file means it'll pop out with a 20B base size.
+    if (node.buffer.length) {
+      gzipSize(node.buffer, function (err, size) {
+        if (err) data.size = node.buffer.length
+        else data.size = size
+        render()
+      })
+    }
+    render()
+  })
+
+  compiler.on('sse-connect', render)
+  compiler.on('sse-disconnect', render)
 
   process.stdout.on('resize', onresize)
 
@@ -211,4 +264,17 @@ function spaceBetween (left, right) {
     space += ' '
   }
   return left + space + right
+}
+
+function time () {
+  var date = new Date()
+  var hours = numPad(date.getHours())
+  var minutes = numPad(date.getMinutes())
+  var seconds = numPad(date.getSeconds())
+  return `${hours}:${minutes}:${seconds}`
+}
+
+function numPad (num) {
+  if (num < 10) num = '0' + num
+  return num
 }
